% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{annotateArms}
\alias{annotateArms}
\title{Annotates measures of CIN}
\usage{
annotateArms(VAULT)
}
\arguments{
\item{VAULT}{The vault object}
}
\value{
Annotated chr arm gains/losses
}
\description{
Annotates measures of CIN
}
