% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{annotateFusions}
\alias{annotateFusions}
\title{Annotates fusions}
\usage{
annotateFusions(VAULT)
}
\arguments{
\item{VAULT}{The vault object}
}
\value{
The annotated fusions with ENSG/gene names
}
\description{
Annotates fusions
}
