% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{annotateGeneCopyNumber}
\alias{annotateGeneCopyNumber}
\title{Given the vault, gets the average basepair-weighted C value for each chromosome arm.}
\usage{
annotateGeneCopyNumber(VAULT)
}
\arguments{
\item{VAULT}{The vault object}
}
\value{
Data table with gene, CN annotation, and required information
}
\description{
Given the vault, gets the average basepair-weighted C value for each chromosome arm.
}
