% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{annotateGeneExpression}
\alias{annotateGeneExpression}
\title{Annotates measures of CIN}
\usage{
annotateGeneExpression(VAULT, GOBJ, OPTS)
}
\arguments{
\item{VAULT}{The vault object}

\item{GOBJ}{the cnvex Gobj via getGobj()}

\item{OPTS}{cnvex opts via getOpts()}
}
\value{
The segments
}
\description{
Annotates measures of CIN
}
