% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaVault_annotate.R
\name{annotateMetaVault}
\alias{annotateMetaVault}
\title{Given the metavault, annotates.}
\usage{
annotateMetaVault(METAVAULT, GOBJ, OPTS)
}
\arguments{
\item{METAVAULT}{The metavault}

\item{GOBJ}{The gobj}

\item{OPTS}{The opts}
}
\value{
list of annotated data.table
}
\description{
Given the metavault, annotates.
}
