% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{annotateMutations}
\alias{annotateMutations}
\title{Annotates the mutations for provided type}
\usage{
annotateMutations(VAULT, TYPE)
}
\arguments{
\item{VAULT}{The vault object}

\item{TYPE}{somatic/germline}
}
\value{
The mutations
}
\description{
Annotates the mutations for provided type
}
