% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{annotateMutationsStructural}
\alias{annotateMutationsStructural}
\title{Annotates the structural mutations for provided type}
\usage{
annotateMutationsStructural(VAULT)
}
\arguments{
\item{VAULT}{The vault object}

\item{TYPE}{structural}
}
\value{
The mutations

TODO: this currently oversimplifies and misses variants in different alleles
}
\description{
Annotates the structural mutations for provided type
}
