% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{annotateVault}
\alias{annotateVault}
\title{Annotates vault}
\usage{
annotateVault(VAULT, GOBJ, OPTS)
}
\arguments{
\item{VAULT}{The vault object}

\item{GOBJ}{the cnvex Gobj via getGobj()}

\item{OPTS}{cnvex opts via getOpts()}
}
\value{
List of annotated data
}
\description{
Annotates vault
}
