% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_documentation.R
\name{api_documentation}
\alias{api_documentation}
\title{Browse API Documentation}
\usage{
api_documentation()
}
\value{
\code{NULL}, invisibly.
}
\description{
This function opens in a web browser the documentation of the Rust API.
}
