% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors.R
\name{authors}
\alias{authors}
\title{Identify Authorship of Rust Crates}
\usage{
authors()
}
\value{
\code{NULL}, invisibly.
}
\description{
Since depending crates are vendored by the \code{\link{build_for_cran}} function, the authorship and copyright must be
declared in the \code{DESCRIPTION} file prior to building the source package
for \href{https://cran.r-project.org/}{The Comprehensive R Archive Network (CRAN)}.  This function helps to identify these attributions but is not
guaranteed to the exhaustive, so manual inspection in warranted before submitting to CRAN.
}
