% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf.R
\name{calculateVaultCCF}
\alias{calculateVaultCCF}
\title{Calculates (meta)vault CCF.}
\usage{
calculateVaultCCF(DTV, CFG, FIELD = "somatic")
}
\arguments{
\item{DTV}{dtVault}

\item{CFG}{cargo config via cargoConfig()}

\item{FIELD}{which table to calculate for; default somatic}
}
\value{
vault with CCF/M in selected table
}
\description{
Calculates (meta)vault CCF.
}
