% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.R
\name{circosPlot}
\alias{circosPlot}
\title{make circos plots of the CNV, SV and Fusion data, as available}
\usage{
circosPlot(v, savef = TRUE, gnm = "hg38")
}
\arguments{
\item{savef}{\link{boolean} Save the pdf}

\item{gnm}{genome \link{hg38,mm10}}

\item{mv}{a TPO dtVault}
}
\value{
if save==FALSE, return the circos plot to the default graphics device, if save==TRUE, return nothing
}
\description{
make circos plots of the CNV, SV and Fusion data, as available
}
