% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.R
\name{cohortCicosPlot}
\alias{cohortCicosPlot}
\title{make circos plots of the CNV, SV and Fusion data, summarizing recurrent events}
\usage{
cohortCicosPlot(
  mv,
  bin_size = "cytoband",
  file = NULL,
  title = NULL,
  cnv = TRUE,
  sv = TRUE,
  fusion = TRUE,
  gnm = "hg38"
)
}
\arguments{
\item{mv}{a TPO dtVault}

\item{bin_size}{currently only supports 'cytoband'}

\item{file}{\link{NULL} pdf to save, if NULL will use default graphics device}

\item{cnv, sv, fusion}{booleans for which data types to display \link{Default: TRUE}}
}
\value{
if save==FALSE, return the circos plot to the default graphics device, if save==TRUE, return nothing
}
\description{
make circos plots of the CNV, SV and Fusion data, summarizing recurrent events
}
