% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depracated.R
\name{combineMapsTables}
\alias{combineMapsTables}
\title{Combines tables from lists of vaults. Wrapper fxn.}
\usage{
combineMapsTables(VAULTS, FIELD1, FIELD2)
}
\arguments{
\item{VAULTS}{A list of filtered vault object}

\item{FIELD1}{Table/map of vault}

\item{FIELD2}{Which table/map}
}
\value{
data.table of combined TABLE with id ID
}
\description{
Combines tables from lists of vaults. Wrapper fxn.
}
