% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depracated.R
\name{combineMeta}
\alias{combineMeta}
\title{Creates data.table from metadata of a list of vaults}
\usage{
combineMeta(VAULTS)
}
\arguments{
\item{VAULTS}{List of vaults}
}
\value{
data.table of meta data from all vaults
}
\description{
Creates data.table from metadata of a list of vaults
}
