% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{compressOverlapSegments}
\alias{compressOverlapSegments}
\title{Given overlap returns the compressed overlap}
\usage{
compressOverlapSegments(OVERLAP)
}
\arguments{
\item{OVERLAP}{dataframe of overlap of segments with arm}
}
\value{
The compressed overlap df
}
\description{
Given overlap returns the compressed overlap
}
