% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{createDuckDBFromVault}
\alias{createDuckDBFromVault}
\title{Create DuckDB from a legacy vault or dtVault}
\usage{
createDuckDBFromVault(dbfn, vault, anno = NULL)
}
\arguments{
\item{dbfn}{DuckDB filename}

\item{vault}{legacy or dtVault}

\item{anno}{dtVaultAnno object}
}
\description{
Create DuckDB from a legacy vault or dtVault
}
