% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{createDuckDBFromVaultFiles}
\alias{createDuckDBFromVaultFiles}
\title{Create DuckDB from legacy vault or dtVault RDS files}
\usage{
createDuckDBFromVaultFiles(dbfn, rds_fns, anno = NULL, verbose = TRUE)
}
\arguments{
\item{dbfn}{DuckDB filename}

\item{rds_fns}{Vault RDS files}

\item{anno}{dtVaultAnno}

\item{verbose}{display progress}
}
\description{
Create DuckDB from legacy vault or dtVault RDS files
}
