% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoprint.R
\name{createOncoArms}
\alias{createOncoArms}
\title{Returns a df usable for making oncoMatrix}
\usage{
createOncoArms(ANNOTATED, ARMS)
}
\arguments{
\item{ANNOTATED}{An annotated metavault}

\item{FIELD}{Part of annotated metavault}
}
\value{
A matrix of chr arm gain/loss
}
\description{
Returns a df usable for making oncoMatrix
}
