% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoprint.R
\name{createOncoCin}
\alias{createOncoCin}
\title{Returns a df usable for making oncoMatrix}
\usage{
createOncoCin(ANNOTATED, FIELD)
}
\arguments{
\item{ANNOTATED}{An annotated metavault}

\item{FIELD}{Part of annotated metavault}
}
\value{
A matrix of CIN values
}
\description{
Returns a df usable for making oncoMatrix
}
