% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoprint.R
\name{createOncoMatrix}
\alias{createOncoMatrix}
\title{Returns a matrix usable for oncoprints}
\usage{
createOncoMatrix(FEATURES, ANNOTATED, FIELD, SUBFIELD)
}
\arguments{
\item{FEATURES}{A vector of gene names}

\item{SAMPLES}{A vector of sample IDs}

\item{HITS}{A 3 col dataframe of hits}
}
\value{
A df usable to make oncoprint
}
\description{
Returns a matrix usable for oncoprints
}
