% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbVault.R
\name{dbVault}
\alias{dbVault}
\title{Create dbVault object from a DB connection}
\usage{
dbVault(con)
}
\arguments{
\item{con}{Supported DB connection}
}
\value{
dbVault
}
\description{
Create dbVault object from a DB connection
}
