% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbVault.R
\name{dbVaultCollect}
\alias{dbVaultCollect}
\title{Collect on-disk dbVault to in-memory dtVault}
\usage{
dbVaultCollect(dbv, index = TRUE, lazy = TRUE)
}
\arguments{
\item{dbv}{dbVault}

\item{index}{boolean; whether to restore dtVault indexes}
}
\value{
dtVault
}
\description{
Collect on-disk dbVault to in-memory dtVault
}
