% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{.annotateArms}
\alias{.annotateArms}
\title{Given the arm returns, compresses region, overlaps with segments and returns
the annotation.}
\usage{
.annotateArms(ARM, VAULT, POS, WM)
}
\arguments{
\item{ARM}{the arm you're annotating}

\item{WM}{The sample's weighted median copy number}

\item{DIGEST}{The digest file}

\item{ARMS}{The coordinates for the arms}
}
\value{
The annotated unified object.
}
\description{
Given the arm returns, compresses region, overlaps with segments and returns
the annotation.
}
