% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{.annotateFusions}
\alias{.annotateFusions}
\title{Annotates fusions for specified field}
\usage{
.annotateFusions(INT, FIELD)
}
\arguments{
\item{FIELD}{The portion of the fusion to reformat}

\item{VAULT}{The vault object}
}
\value{
The annotated fusions for specified field
}
\description{
Annotates fusions for specified field
}
