% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf.R
\name{.calcM}
\alias{.calcM}
\title{Calculates multiplicity of variant}
\usage{
.calcM(DL, CFG)
}
\arguments{
\item{DL}{data list}

\item{CFG}{cargo config via cargoConfig()}
}
\value{
vector of Ms - multiplicities
}
\description{
Calculates multiplicity of variant
}
