% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf.R
\name{.calculateVaultCCF}
\alias{.calculateVaultCCF}
\title{Calculates vault CCF for an individual vault}
\usage{
.calculateVaultCCF(V, CFG, FIELD = "somatic")
}
\arguments{
\item{V}{The vault}

\item{CFG}{cargo config via cargoConfig()}

\item{FIELD}{which table to calculate for; default somatic}
}
\value{
vault with CCF/M in selected table
}
\description{
Calculates vault CCF for an individual vault
}
