% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triage.R
\name{.dxRescueCheck}
\alias{.dxRescueCheck}
\title{Returns id--var_id pairs from teh table that are not present in the vault}
\usage{
.dxRescueCheck(V, TBL)
}
\arguments{
\item{V}{\itemize{
\item vault
}}

\item{TBL}{a table containing id,var_id,table,& pass value}
}
\value{
vector of id-var_id pairs not present in the vault but in the table
}
\description{
Returns id--var_id pairs from teh table that are not present in the vault
}
