% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triage.R
\name{.dxRescueVariants}
\alias{.dxRescueVariants}
\title{Changes pass/fail value of a vault variant table based on provided values}
\usage{
.dxRescueVariants(NEW_VALS, VAULT_TABLE)
}
\arguments{
\item{NEW_VALS}{\itemize{
\item the variant table with new values
}}

\item{VAULT_TABLE}{the vault table needing to be updated w/ new values}
}
\value{
updated vault
}
\description{
Changes pass/fail value of a vault variant table based on provided values
}
