% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dxVault.R
\name{.dxVaultSubset}
\alias{.dxVaultSubset}
\title{Subset dxVault by column and value}
\usage{
.dxVaultSubset(dxv, filter.col, filter.val)
}
\arguments{
\item{dxv}{dbVault dpVault dtVault}

\item{filter.col}{column to filter on}

\item{filter.val}{values to keep}
}
\value{
dxVault
}
\description{
If a column is not present in a table it is a no-op
}
