% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{.fusionFormat}
\alias{.fusionFormat}
\title{reformats fusion table for easier viewing;
slightly modified version of svFormat}
\usage{
.fusionFormat(TBL, REORDER = TRUE)
}
\arguments{
\item{TBL}{fusion table}

\item{REORDER}{should the table be reordered based on 'chain'}
}
\value{
A user-friendly fusion table
}
\description{
reformats fusion table for easier viewing;
slightly modified version of svFormat
}
