% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{.getMutations}
\alias{.getMutations}
\title{gets id/GENE key}
\usage{
.getMutations(TBL)
}
\arguments{
\item{TABLE}{which table to get}
}
\value{
vector of id--gene values
}
\description{
gets id/GENE key
}
