% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{.labelCN}
\alias{.labelCN}
\title{Annotates copy number of a row based on
C,K,sex, and chromosome arm average C value}
\usage{
.labelCN(ROW, SEX)
}
\arguments{
\item{ROW}{The unified output corresponding to a single gene}

\item{SEX}{Sex of sample from vault object}
}
\value{
The annotation for the gene
}
\description{
Annotates copy number of a row based on
C,K,sex, and chromosome arm average C value
}
