% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depracated.R
\name{.makeMapFromTable}
\alias{.makeMapFromTable}
\title{Given a table, makes a map}
\usage{
.makeMapFromTable(TBL)
}
\arguments{
\item{TBL}{vault table}
}
\value{
vault map
}
\description{
Given a table, makes a map
}
