% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.parseGtfAttribute}
\alias{.parseGtfAttribute}
\title{Parses GTF attribute}
\usage{
.parseGtfAttribute(ATTRIBUTE, GTF)
}
\arguments{
\item{ATTRIBUTE}{\itemize{
\item the attribute to parse
}}

\item{GTF}{path to gtf file, default currently is my file}
}
\value{
GTF object
}
\description{
Parses GTF attribute
}
