% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metaVault.R
\name{.plotCnHeatmapActualC}
\alias{.plotCnHeatmapActualC}
\title{Plots actual C (unadjusted C)}
\usage{
.plotCnHeatmapActualC(ANNOTATED, META)
}
\arguments{
\item{ANNOTATED}{Output of prepareCnHeatmap}

\item{META}{optional DF of meta to order/split. 2 columns ('order'/'split') with ids as rownames.}
}
\value{
Actual C plot
}
\description{
Plots actual C (unadjusted C)
}
