% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metaVault.R
\name{.plotCnHeatmapAnnotatedC}
\alias{.plotCnHeatmapAnnotatedC}
\title{Plots Annotated heatmap (amplification, gain, ect.)}
\usage{
.plotCnHeatmapAnnotatedC(ANNOTATED, META)
}
\arguments{
\item{ANNOTATED}{Output of prepareCnHeatmap}

\item{META}{optional DF of meta to order/split. 2 columns ('order'/'split') with ids as rownames.}
}
\value{
Annotated plot
}
\description{
Plots Annotated heatmap (amplification, gain, ect.)
}
