% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metaVault.R
\name{.plotCnHeatmapBinaryK}
\alias{.plotCnHeatmapBinaryK}
\title{Plots binary K (LOH)}
\usage{
.plotCnHeatmapBinaryK(ANNOTATED, META)
}
\arguments{
\item{ANNOTATED}{Output of prepareCnHeatmap}

\item{META}{optional DF of meta to order/split. 2 columns ('order'/'split') with ids as rownames.}
}
\value{
Binary K plot
}
\description{
Plots binary K (LOH)
}
