% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metaVault.R
\name{.plotCnHeatmapRelativeC}
\alias{.plotCnHeatmapRelativeC}
\title{Plots relative C (C:ploidy)}
\usage{
.plotCnHeatmapRelativeC(ANNOTATED, META)
}
\arguments{
\item{ANNOTATED}{Output of prepareCnHeatmap}

\item{META}{optional DF of meta to order/split. 2 columns ('order'/'split') with ids as rownames.}
}
\value{
Relative C plot
}
\description{
Plots relative C (C:ploidy)
}
