% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoprint.R
\name{.prepareOncoHits}
\alias{.prepareOncoHits}
\title{Returns a df usable for making oncoMatrix}
\usage{
.prepareOncoHits(ANNOTATED, FIELD, SUBFIELD, FEATURES)
}
\arguments{
\item{ANNOTATED}{An annotated metavault}

\item{FIELD}{Part of annotated metavault}

\item{SUBFIELD}{Type of FIELD (if necessary)}

\item{FEATURES}{the features of interest}
}
\value{
A df of hits of specified type
}
\description{
Returns a df usable for making oncoMatrix
}
