% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoprint.R
\name{.prepareOncoMissing}
\alias{.prepareOncoMissing}
\title{Returns a df usable for making oncoMatrix}
\usage{
.prepareOncoMissing(ANNOTATED, FIELD, FEATURES)
}
\arguments{
\item{ANNOTATED}{An annotated metavault}

\item{FIELD}{Part of annotated metavault}

\item{FEATURES}{the features of interest}
}
\value{
A df of hits of specified type
}
\description{
Returns a df usable for making oncoMatrix
}
