% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf.R
\name{.prepareVaultCCF}
\alias{.prepareVaultCCF}
\title{Prepares mutation table for CCF calculation. Robust to MV}
\usage{
.prepareVaultCCF(V, FIELD)
}
\arguments{
\item{V}{The vault}

\item{FIELD}{which table}
}
\value{
prepared data table
}
\description{
Prepares mutation table for CCF calculation. Robust to MV
}
