% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistic.R
\name{.prepGisticTable}
\alias{.prepGisticTable}
\title{Prep gistic table for single sample}
\usage{
.prepGisticTable(ID, TBL, SEG_CN_METHOD)
}
\arguments{
\item{ID}{The sample ID}

\item{TBL}{The segment table}

\item{SEG_CN_METHOD}{How the seg.CN parameter is calculated}
}
\value{
Table for gistic for a single sample
}
\description{
Prep gistic table for single sample
}
