% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depracated.R
\name{.removeFailedVariantTable}
\alias{.removeFailedVariantTable}
\title{Removes variants that do not pass filters from a specific table}
\usage{
.removeFailedVariantTable(V, TABLE)
}
\arguments{
\item{V}{\itemize{
\item filtered vault
}}

\item{TABLE}{\itemize{
\item the table to remove failed variants
}}
}
\value{
updated vault
}
\description{
Removes variants that do not pass filters from a specific table
}
