% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{.reviewFeaturesExpression}
\alias{.reviewFeaturesExpression}
\title{Compiles expression data}
\usage{
.reviewFeaturesExpression(MVF, FEATURES)
}
\arguments{
\item{MVF}{filtered metavault}

\item{FEATURES}{vector of features; gene_names}
}
\value{
trpkm for features
}
\description{
Compiles expression data
}
