% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{.reviewFeaturesGeneCopy}
\alias{.reviewFeaturesGeneCopy}
\title{Compiles variants}
\usage{
.reviewFeaturesGeneCopy(MVF, FEATURES)
}
\arguments{
\item{MVF}{filtered metavault}

\item{FEATURES}{vector of features; gene_names}
}
\value{
A table of pertinent variant cols
}
\description{
Compiles variants
}
