% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistic.R
\name{.segCN}
\alias{.segCN}
\title{Compute seg.CN}
\usage{
.segCN(SEGMENTS, SEG_CN_METHOD = "centered_c")
}
\arguments{
\item{SEGMENTS}{The vault segment table}

\item{SEG_CN_METHOD}{How the seg.CN parameter is calculated}
}
\value{
vector of seg.CN values
}
\description{
Compute seg.CN
}
