% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{.tallyChrAlterations}
\alias{.tallyChrAlterations}
\title{Tallies chromosomes gained/lost in several ways}
\usage{
.tallyChrAlterations(ARMS, VALUE)
}
\arguments{
\item{ARMS}{The output of annotateArms()}

\item{VALUE}{which tally to return}
}
\value{
integer of total number of whole chromosomes gained/lost in defined way
}
\description{
Tallies chromosomes gained/lost in several ways
}
