% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{.tallyCN}
\alias{.tallyCN}
\title{tallies copy number hits}
\usage{
.tallyCN(HITS, GC)
}
\arguments{
\item{HITS}{the hits table}

\item{GC}{gene copy numebr table}
}
\value{
vector of cn losses
}
\description{
tallies copy number hits
}
