% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{.tallyMutations}
\alias{.tallyMutations}
\title{tallies mutations}
\usage{
.tallyMutations(HITS, TBL)
}
\arguments{
\item{HITS}{the hits table}

\item{TBL}{the mutation type}
}
\value{
table
}
\description{
tallies mutations
}
