% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depracated.R
\name{.triageAnnotationVault}
\alias{.triageAnnotationVault}
\title{Changes pass/fail value of variants based on annotation filter}
\usage{
.triageAnnotationVault(vault)
}
\arguments{
\item{vault}{\itemize{
\item vault
}}

\item{anno}{anno object (list); contains gene info}
}
\value{
filtered vault
}
\description{
Changes pass/fail value of variants based on annotation filter
}
