% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depracated.R
\name{.triageArtifactVault}
\alias{.triageArtifactVault}
\title{Changes pass/fail value of variants based on artifact filter}
\usage{
.triageArtifactVault(vault)
}
\arguments{
\item{vault}{\itemize{
\item vault
}}
}
\value{
filtered vault
}
\description{
Changes pass/fail value of variants based on artifact filter
}
