% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depracated.R
\name{.triageConsequenceVault}
\alias{.triageConsequenceVault}
\title{Changes pass/fail value of variants based on consequence filter}
\usage{
.triageConsequenceVault(vault)
}
\arguments{
\item{vault}{\itemize{
\item vault
}}
}
\value{
filtered vault
}
\description{
Changes pass/fail value of variants based on consequence filter
}
