% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depracated.R
\name{.triageEvidenceVault}
\alias{.triageEvidenceVault}
\title{Changes pass/fail value of variants based on evidence filter}
\usage{
.triageEvidenceVault(vault, GOI = NULL)
}
\arguments{
\item{vault}{\itemize{
\item vault
}}

\item{GOI}{\itemize{
\item optional vector of genes of interst
}}
}
\value{
filtered vault
}
\description{
Changes pass/fail value of variants based on evidence filter
}
