% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triage.R
\name{.triageFusion}
\alias{.triageFusion}
\title{Determines triage status of gene fusions}
\usage{
.triageFusion(fusion, anno)
}
\arguments{
\item{fusion}{fusion table}

\item{anno}{anno table}
}
\value{
triaged table
}
\description{
Determines triage status of gene fusions
}
