% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triage.R
\name{.triageGermline}
\alias{.triageGermline}
\title{Determines triage status of germline variants}
\usage{
.triageGermline(germline, anno)
}
\arguments{
\item{germline}{germline table}

\item{anno}{anno table}
}
\value{
filtered vault
}
\description{
Determines triage status of germline variants
}
