% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depracated.R
\name{.triagePhaseVault}
\alias{.triagePhaseVault}
\title{Changes pass/fail value of variants based on phase-id (pid) filter}
\usage{
.triagePhaseVault(vault)
}
\arguments{
\item{vault}{\itemize{
\item vault
}}
}
\value{
filtered vault
}
\description{
Changes pass/fail value of variants based on phase-id (pid) filter
}
