% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triage.R
\name{.triageSomatic}
\alias{.triageSomatic}
\title{Determines triage status of somatic variants}
\usage{
.triageSomatic(somatic, anno)
}
\arguments{
\item{somatic}{somatic table}

\item{anno}{anno table}
}
\value{
triaged table
}
\description{
Determines triage status of somatic variants
}
