% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triage.R
\name{.triageStructural}
\alias{.triageStructural}
\title{Determines triage status of structural variants}
\usage{
.triageStructural(structural, anno)
}
\arguments{
\item{structural}{structural table}

\item{anno}{anno table}
}
\value{
triaged table
}
\description{
Determines triage status of structural variants
}
