% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpVault.R
\name{dpVaultCollect}
\alias{dpVaultCollect}
\title{Collect dpVault to dtVault}
\usage{
dpVaultCollect(dpv, index = TRUE)
}
\arguments{
\item{dpv}{dpVault}

\item{index}{boolean; whether to restore dtVault indexes}
}
\value{
dtVault
}
\description{
Collect dpVault to dtVault
}
