% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtVault.R
\name{dtVault}
\alias{dtVault}
\title{Create dtVault object from legacy vault or metavault}
\usage{
dtVault(vault, anno = NULL, gene_model = NULL, ...)
}
\arguments{
\item{vault}{legacy vault}

\item{anno}{dtVaultAnno object}

\item{gene_model}{GTF file}
}
\value{
dtVault
}
\description{
Create dtVault object from legacy vault or metavault
}
