% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtVault.R
\name{dtVaultAnno}
\alias{dtVaultAnno}
\title{Create dtVaultAnno object}
\usage{
dtVaultAnno(
  gene_model,
  somatic_goi = NULL,
  germline_goi = NULL,
  somatic_art = NULL,
  drop_chrM = TRUE
)
}
\arguments{
\item{gene_model}{GTF file}

\item{somatic_goi}{custom list of somatic genes of interest \link{default: goi-somatic-tier1}}

\item{germline_goi}{custom list of germline genes of interest \link{default: goi-germline}}

\item{somatic_art}{custom list of somatic artifact genes \link{default: art-somatic}}

\item{drop_chrM}{remove genes on chrM}
}
\value{
dtVaultAnno
}
\description{
Create dtVaultAnno object
}
