% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtVault.R
\name{dtVaultIndex}
\alias{dtVaultIndex}
\title{Restore dtVault default indexes}
\usage{
dtVaultIndex(dtv)
}
\arguments{
\item{dbv}{dtVault}
}
\value{
dtVault
}
\description{
Note: this opperation occurs in-place
}
