% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtVault.R
\name{dtVaultStack}
\alias{dtVaultStack}
\title{Stack (rbind) list of dtVaults}
\usage{
dtVaultStack(vaults, anno = NULL)
}
\arguments{
\item{vaults}{list of dtVaults}

\item{anno}{dtAnno object \link{optional}}
}
\value{
dtVault
}
\description{
Stack (rbind) list of dtVaults
}
