% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtVault.R
\name{dtVaultUpdate}
\alias{dtVaultUpdate}
\title{Update vault to newer dtVault format}
\usage{
dtVaultUpdate(vault, anno = NULL, format = "v2")
}
\arguments{
\item{vault}{legacy vault}

\item{anno}{dtVaultAnno object}

\item{format}{dtVault format \link{v2,v3}}
}
\value{
dtVault
}
\description{
Note: input vault is modified
}
