% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triage.R
\name{dxRescueVariants}
\alias{dxRescueVariants}
\title{Changes triage value of variants based on provided table}
\usage{
dxRescueVariants(V, TBL)
}
\arguments{
\item{V}{\itemize{
\item vault
}}

\item{TBL}{a table containing id,var_id,table,& pass value}
}
\value{
updated vault
}
\description{
Changes triage value of variants based on provided table
}
