% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dxVault.R
\name{dxVaultApply}
\alias{dxVaultApply}
\title{Apply function to dxVault subsets}
\usage{
dxVaultApply(dxv, fun, split = "id")
}
\arguments{
\item{dxv}{dbVault dpVault dtVault}

\item{fun}{function operating on vault}

\item{split}{subset by 'id' or 'gene'}
}
\value{
dxVault
}
\description{
Apply function to dxVault subsets
}
