% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dxVault.R
\name{dxVaultCheck}
\alias{dxVaultCheck}
\title{Check if vault has allowed storage and version}
\usage{
dxVaultCheck(
  dxv,
  storage = c("data.table", "duckdb", "tbl_lazy"),
  format = c("v2")
)
}
\arguments{
\item{dxv}{dbVault dpVault dtVault}

\item{storage}{allowed storage type \link{data.table,duckdb,tbl_lazy}}

\item{format}{allowed format version \link{v2}}
}
\description{
By default allows all backends
}
