% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dxVault.R
\name{dxVaultCheckCardinality}
\alias{dxVaultCheckCardinality}
\title{Check if vault has allowed storage and version}
\usage{
dxVaultCheckCardinality(dxv, expect = 1)
}
\arguments{
\item{dxv}{dbVault dpVault dtVault}

\item{expect}{expected cardinality default: 1}
}
\description{
Check if vault has allowed storage and version
}
