% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dxVault.R
\name{dxVaultCollect}
\alias{dxVaultCollect}
\title{Collect on-disk dbVault or in-memory dpVault}
\usage{
dxVaultCollect(dxv, ...)
}
\arguments{
\item{...}{pass-through parameters to dbVaultCollect and dpVaultCollect}

\item{dbv}{dbVault}
}
\value{
dtVault
}
\description{
Collect on-disk dbVault or in-memory dpVault
}
