% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dxVault.R
\name{dxVaultSubsetByGene}
\alias{dxVaultSubsetByGene}
\title{Subset dxVault by gene}
\usage{
dxVaultSubsetByGene(dxv, gene_names = NULL, gene_ids = NULL)
}
\arguments{
\item{dxv}{dbVault dpVault dtVault}

\item{gene_names}{one or multiple gene symbols}

\item{gene_ids}{one or multipe gene ids}
}
\value{
dxVault
}
\description{
Subset dxVault by gene
}
