% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dxVault.R
\name{dxVaultSubsetById}
\alias{dxVaultSubsetById}
\title{Subset dxVault by id}
\usage{
dxVaultSubsetById(dxv, ids)
}
\arguments{
\item{dxv}{dbVault dpVault dtVault}

\item{ids}{one or multipe ids}
}
\value{
dxVault
}
\description{
Subset dxVault by id
}
