% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dxVault.R
\name{dxVaultSubsetByIndex}
\alias{dxVaultSubsetByIndex}
\title{Subset dxVault by index}
\usage{
dxVaultSubsetByIndex(dxv, idxs)
}
\arguments{
\item{dxv}{dbVault dpVault dtVault}

\item{idxs}{one or multipe indexes}
}
\value{
dxVault
}
\description{
Subset dxVault by index
}
