% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dxVault.R
\name{dxVaultSubsetByTable}
\alias{dxVaultSubsetByTable}
\title{Subset dxVault by table}
\usage{
dxVaultSubsetByTable(dxv, tables)
}
\arguments{
\item{dxv}{dbVault dpVault dtVault}

\item{tables}{one or multipe table names}
}
\value{
dxVault
}
\description{
Subset dxVault by table
}
