% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf.R
\name{estimatePurityFromSomaticVariants}
\alias{estimatePurityFromSomaticVariants}
\title{Estimate purity from variants and CNV model}
\usage{
estimatePurityFromSomaticVariants(dtv, strategy = "diploid-modal")
}
\arguments{
\item{dtv}{dtVault (1 sample)}

\item{strategy}{named heuristic used to estimate purity \link{diploid-modal}}
}
\value{
purity (numeric)
}
\description{
Estimate purity from variants and CNV model
}
