% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand}
\alias{expand}
\title{Run Cargo Expand for R Packages}
\usage{
expand()
}
\value{
\code{NULL}, invisibly.
}
\description{
The 'roxido' framework uses macros and this function will expand all macros and display to source code.
}
