% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depracated.R
\name{extractGene}
\alias{extractGene}
\title{Subsets a metavault to only contain select genes by id}
\usage{
extractGene(METAVAULT, GENEIDS = NULL, SYMBOLS = NULL, ANNO = NULL)
}
\arguments{
\item{METAVAULT}{metaVault}

\item{GENEIDS}{vector of gene ids}

\item{SYMBOLS}{vector of gene symbols (optional)}

\item{ANNO}{vault annotation object}
}
\value{
vault for select ids
}
\description{
Subsets a metavault to only contain select genes by id
}
