% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depracated.R
\name{extractVault}
\alias{extractVault}
\title{Subsets vault(s) from a metavault}
\usage{
extractVault(IDENTIFIERS, METAVAULT)
}
\arguments{
\item{METAVAULT}{A metaVault}

\item{IDENTIFIER}{A ID within the metaVault}
}
\value{
vault for that id
}
\description{
Subsets vault(s) from a metavault
}
