% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depracated.R
\name{filterVault}
\alias{filterVault}
\title{Filter variants based on "pass" filter}
\usage{
filterVault(V)
}
\arguments{
\item{V}{\itemize{
\item triaged vault
}}
}
\value{
vault with failed variants (somatic, germline, structural, fusion) removed
}
\description{
Filter variants based on "pass" filter
}
