% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{formatReviewTable}
\alias{formatReviewTable}
\title{reformats review table for storage}
\usage{
formatReviewTable(TBL, FIELD)
}
\arguments{
\item{TBL}{output of getReviewTable fxns; post manual review}

\item{FIELD}{which table the variant originates from}
}
\value{
A table that can be used with
}
\description{
reformats review table for storage
}
