% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaVault_annotate.R
\name{getAnnotListTable}
\alias{getAnnotListTable}
\title{Gets tables from annotated list}
\usage{
getAnnotListTable(ALIST, FIELD)
}
\arguments{
\item{ALIST}{The annotated list}

\item{FIELD}{the table to get}
}
\value{
list of annotated data.table
}
\description{
Gets tables from annotated list
}
