% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{getArmAverage}
\alias{getArmAverage}
\title{Given the vault, gets the average basepair-weighted C value for each chromosome arm.}
\usage{
getArmAverage(VAULT)
}
\arguments{
\item{VAULT}{The vault object}
}
\value{
data.table with gene and arm average
}
\description{
Given the vault, gets the average basepair-weighted C value for each chromosome arm.
}
