% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{getArmInfo}
\alias{getArmInfo}
\title{Pulls the coordinates for chromosome arms given an assembly}
\usage{
getArmInfo(ASSEMBLY)
}
\arguments{
\item{ASSEMBLY}{The vault's assembly information}
}
\value{
A GRanges object with the chromosome arm information
TODO: Handle other assemblies
}
\description{
Pulls the coordinates for chromosome arms given an assembly
}
