% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{getArms}
\alias{getArms}
\title{gets the chr arms for the vault's assembly}
\usage{
getArms(VAULT)
}
\arguments{
\item{VAULT}{The vault object}
}
\value{
df of chr arms

@export
}
\description{
gets the chr arms for the vault's assembly
}
