% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{getGeneSegLength}
\alias{getGeneSegLength}
\title{Gets weighted median of sample.}
\usage{
getGeneSegLength(VAULT)
}
\arguments{
\item{VAULT}{The vault object}
}
\value{
the weighted median of the library
TODO: This might be leaking
}
\description{
Gets weighted median of sample.
}
