% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metaVault.R
\name{getOverlapWidth}
\alias{getOverlapWidth}
\title{Given 2 granges objects, returns the overlap w/ widths of overlap}
\usage{
getOverlapWidth(QUERY, SUBJECT)
}
\arguments{
\item{QUERY}{typically the tiles}

\item{SUBJECT}{typically the segment data}
}
\value{
overlap df with widths of overlap
}
\description{
Given 2 granges objects, returns the overlap w/ widths of overlap
}
