% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{getReviewTableStructural}
\alias{getReviewTableStructural}
\title{generates table for review; subsets cols from structural table & adds from meta}
\usage{
getReviewTableStructural(MV, GENE_NAME)
}
\arguments{
\item{MV}{A metavault}

\item{GENE_NAME}{single gene name}
}
\value{
A table for manual review of variants
}
\description{
generates table for review; subsets cols from structural table & adds from meta
}
