% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metaVault.R
\name{getSegments}
\alias{getSegments}
\title{Returns segments for all samples that have some non-NA C values}
\usage{
getSegments(MV)
}
\arguments{
\item{MV}{the metavault}
}
\value{
granges of segments
}
\description{
Returns segments for all samples that have some non-NA C values
}
