% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metaVault.R
\name{getSex}
\alias{getSex}
\title{Given meta vault get sex for each id}
\usage{
getSex(MV)
}
\arguments{
\item{MV}{metavault}
}
\value{
df of id/sex
}
\description{
Given meta vault get sex for each id
}
