% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metaVault.R
\name{getTiles}
\alias{getTiles}
\title{Gets tiles given GOBJ and bin width}
\usage{
getTiles(GOBJ, BIN)
}
\arguments{
\item{GOBJ}{the Gobj}

\item{BIN}{width of bin to use on genome; default 1MB}
}
\value{
granges of tiles
}
\description{
Gets tiles given GOBJ and bin width
}
