% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{getVaultId}
\alias{getVaultId}
\title{gets id(s) from vault/metavault/annotated metavault}
\usage{
getVaultId(V)
}
\arguments{
\item{V}{The vault-like object}
}
\value{
vector of ids

@export
}
\description{
gets id(s) from vault/metavault/annotated metavault
}
