% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{getVaultMap}
\alias{getVaultMap}
\title{Gets specific vault map}
\usage{
getVaultMap(V, FIELD)
}
\arguments{
\item{V}{vault}

\item{FIELD}{which map}
}
\value{
vault map
}
\description{
Gets specific vault map
}
