% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{getVaultTable}
\alias{getVaultTable}
\title{Gets specific vault table}
\usage{
getVaultTable(V, FIELD)
}
\arguments{
\item{V}{vault}

\item{FIELD}{which table}
}
\value{
vault table
}
\description{
Gets specific vault table
}
