% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{getWeightedMedian}
\alias{getWeightedMedian}
\title{Gets weighted median of sample.}
\usage{
getWeightedMedian(VAULT)
}
\arguments{
\item{VAULT}{The vault object}
}
\value{
the weighted median of the library
}
\description{
Gets weighted median of sample.
}
