% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depracated.R
\name{injectVariants}
\alias{injectVariants}
\title{Inject new variants into the vault}
\usage{
injectVariants(V, TBL, FIELD)
}
\arguments{
\item{V}{a vault}

\item{TBL}{table of variants to add}

\item{FIELD}{which table to add variants to}
}
\value{
updated vault
}
\description{
Inject new variants into the vault
}
