% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargo.R
\name{is_available}
\alias{is_available}
\title{Test Availability of Cargo}
\usage{
is_available(minimum_version, verbose = TRUE)
}
\arguments{
\item{minimum_version}{A character string representing the minimum Rust
version that is needed.}

\item{verbose}{Print information about the installation?}
}
\value{
Logical indicating whether (at least the desired version of) Cargo is
available.
}
\description{
This function checks if Cargo is available on the system.
}
\examples{
if ( ! is_available("1.50") ) {
  cat("Please run 'cargo::install()'.\n")
}

}
