% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{labelCn}
\alias{labelCn}
\title{Wrapper script to annotate cnvex output.}
\usage{
labelCn(INTERMEDIATE, SEX)
}
\arguments{
\item{INTERMEDIATE}{An intermediate object from annotateGeneCopyNumber()}

\item{SEX}{Sex of sample from vault object}
}
\value{
The copy number labels
}
\description{
Wrapper script to annotate cnvex output.
}
