% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{labelOverlap}
\alias{labelOverlap}
\title{Given compressed overlap returns the C, label, and proportion of arm}
\usage{
labelOverlap(OVERLAP)
}
\arguments{
\item{OVERLAP}{dataframe of overlap of segments with arm}
}
\value{
df containing the C values, label, and proportino of arm
}
\description{
Given compressed overlap returns the C, label, and proportion of arm
}
