% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{load_GTF}
\alias{load_GTF}
\title{Loads GTF file from location and adds useful columns}
\usage{
load_GTF(GTF)
}
\arguments{
\item{GTF}{path to gtf file, default currently is my file}
}
\value{
GTF object
}
\description{
Loads GTF file from location and adds useful columns
}
