% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{makeHitsTable}
\alias{makeHitsTable}
\title{Makes a table of hits}
\usage{
makeHitsTable(IDS, GENES)
}
\arguments{
\item{IDS}{vector of ids}

\item{GENES}{DF of GENES and SYMBOLS}
}
\value{
DF of hits
}
\description{
Makes a table of hits
}
