% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depracated.R
\name{metaVault}
\alias{metaVault}
\title{Create metaVault. Main function.}
\usage{
metaVault(VAULTS, ANNO = NULL)
}
\arguments{
\item{VAULTS}{List of vaults}

\item{ANNO}{Object returned by vaultAnno}
}
\value{
metaVault
}
\description{
Create metaVault. Main function.
}
