% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_package.R
\name{new_package}
\alias{new_package}
\title{Make a Skeleton for a New Package}
\usage{
new_package(path, revision = "main", include_justfile = FALSE)
}
\arguments{
\item{path}{A path where the package is created.  The name of the
package is taken as the last element in the file path.}

\item{revision}{A git revision to use.  Defaults to "main".}

\item{include_justfile}{Should the new package have a \code{justfile} for convenience?'}
}
\description{
A new Rust-based package is created at the supplied path and the package is installed.
}
