% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metaVault.R
\name{plotCnHeatmap}
\alias{plotCnHeatmap}
\title{Plots CN heatmaps for samples w/ segment data}
\usage{
plotCnHeatmap(MV, GOBJ, BIN = 1e+06, TYPE, META = NULL)
}
\arguments{
\item{MV}{The metavault}

\item{GOBJ}{the Gobj}

\item{BIN}{width of bin to use on genome; default 1MB}

\item{TYPE}{the type of CN heatmap; c('Annotated', 'RelativeC', 'ActualC')}

\item{META}{optional DF of meta to order/split. 2 columns ('order'/'split') with ids as rownames.}
}
\value{
plot
}
\description{
Plots CN heatmaps for samples w/ segment data
}
