% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metaVault.R
\name{plotLollipop}
\alias{plotLollipop}
\title{Plots Lolliplots from metavaults}
\usage{
plotLollipop(
  MV,
  SYMBOL,
  GTF,
  PROTEIN_DOMAINS = F,
  FORMAT = "MVF",
  PATIENTS = NULL,
  STACK_SAME = "stack",
  REMOVE_INTRON = T,
  FIELD = "somatic",
  PROTEIN = "protein",
  REMOVE_EXCESS = F,
  BREAKPOINT = F,
  COLOR = "rainbow",
  SHOW_EXON_NAME = F,
  INTRON_GAP = NULL,
  LABEL_ON_FEATURE = F,
  SEPARATE = NULL,
  TRANSCRIPT = "common"
)
}
\arguments{
\item{MV}{The metavault or .maf file. If using .maf file then change FORMAT to "MAF"}

\item{SYMBOL}{Gene symbol}

\item{GTF}{GTF table containing exon information}

\item{PROTEIN_DOMAINS}{Add file to plot protein domains on track. Default = F, need to supply protein file as gff3 format.}

\item{FORMAT}{pick input format as metavault "MVF" or MAF format "MAF"}

\item{PATIENTS}{a list of patients to filter by}

\item{STACK_SAME}{How to deal with mutations at same location?; c("stack", "unique", "none")}

\item{REMOVE_INTRON}{Remove intronic regions so exons are easier to see? This will cause issues if there are expected mutations in intron regions!}

\item{PROTEIN}{indicating whether to plot protein or gene info. Choices = c("protein", "gene")}

\item{REMOVE_EXCESS}{Remove exons at either end that contain no information?}

\item{BREAKPOINT}{T/F indicating whether to plot breakpoints based on fusion information. Only works when REMOVE_INTRON is FALSE!}

\item{COLOR}{If you want to add colors to the exons; c("None", "Stripes", "Rainbow")}

\item{INTRON_GAP}{if shrinking introns, how much gap to leave between exons?}

\item{LABEL_ON_FEATURE}{puts exon numbers on the bars instead. But will remove x axis scale.}

\item{SEPARATE}{allows separation of lollipops by up or down if provided a column name. Must be only 2 categories.}

\item{TRANSCRIPT}{Which trascript to use from the GTF. If not specified, it will use the most common transcript found for the gene.}

\item{ZOOM}{If you want to zoom into a specific region to see; c("chrom", "chromstart", "chromend"), else = NULL}
}
\value{
plot
}
\description{
Plots Lolliplots from metavaults
}
