% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistic.R
\name{plotMafGisticChromPlot}
\alias{plotMafGisticChromPlot}
\title{Plot MAF's Chrom plot from GISTIC data}
\usage{
plotMafGisticChromPlot(
  ALL_LESIONS,
  AMP.GENES,
  DEL.GENES,
  SCORES.GISTIC,
  FDR = 0.1,
  REF_BUILD = "hg38",
  CYTOBAND_TEXT_SIZE = 1.3,
  CYTOBAND_OFFSET = 0.05,
  GENES_TEXT_SIZE = 10
)
}
\arguments{
\item{AMP.GENES}{path to amp_genes.conf_99.txt}

\item{DEL.GENES}{path to del_genes.conf_99.txt}

\item{SCORES.GISTIC}{path to scores.gistic}

\item{ALL_LESTIONS}{path to all_lesions.conf_99.txt}
}
\value{
plot
}
\description{
Plot MAF's Chrom plot from GISTIC data
}
