% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-metaVault.R
\name{prepareCnHeatmap}
\alias{prepareCnHeatmap}
\title{prepares data for plotting CN heatmaps}
\usage{
prepareCnHeatmap(MV, GOBJ, BIN = 1e+06)
}
\arguments{
\item{MV}{The metavault}

\item{GOBJ}{the Gobj}

\item{BIN}{width of bin to use on genome; default 1MB}
}
\value{
Df of prepare data
}
\description{
prepares data for plotting CN heatmaps
}
