% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistic.R
\name{prepGisticTable}
\alias{prepGisticTable}
\title{Prep gistic table}
\usage{
prepGisticTable(MV, SEG_CN_METHOD = "centered_c")
}
\arguments{
\item{MV}{The metavault}

\item{SEG_CN_METHOD}{How the seg.CN parameter is calculated}
}
\value{
Table for gistic
}
\description{
Prep gistic table
}
