% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaVault_annotate.R
\name{reformatAnnotVaultList}
\alias{reformatAnnotVaultList}
\title{Given the annotated list, squishes list of lists to list of dfs}
\usage{
reformatAnnotVaultList(ALIST)
}
\arguments{
\item{ALIST}{The annotated list of vaults}
}
\value{
list of annotated data.table
}
\description{
Given the annotated list, squishes list of lists to list of dfs
}
