% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_calls.R
\name{register_calls}
\alias{register_calls}
\title{Generate Rust Code to Register Rust Functions}
\usage{
register_calls(pkgroot = ".")
}
\arguments{
\item{pkgroot}{The root directory of the package.}
}
\value{
\code{NULL}, invisibly.
}
\description{
This function generates Rust code to register Rust functions accessed in R
through .Call(). If a package's usage of .Call() functions changes, rerun
this function to update the \code{src/rustlib/src/registration.rs} file.
}
