% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoprint.R
\name{reorderOncoMatrix}
\alias{reorderOncoMatrix}
\title{Returns a df usable for making oncoMatrix}
\usage{
reorderOncoMatrix(MATRIX, ORDER)
}
\arguments{
\item{MATRIX}{A prepared OncoMatrix}

\item{ORDER}{Vector of IDs in correct order}
}
\value{
A matrix of chr arm gain/loss
}
\description{
Returns a df usable for making oncoMatrix
}
