% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depracated.R
\name{rescueVariant}
\alias{rescueVariant}
\title{Changes pass/fail value of a single variant based on provided values}
\usage{
rescueVariant(V, TABLE, IDD, VAR_IDD, VALUE)
}
\arguments{
\item{V}{\itemize{
\item vault
}}

\item{TABLE}{The table the variant is in}

\item{IDD}{\itemize{
\item The sample ID
}}

\item{VAR_IDD}{\itemize{
\item the variant id
}}

\item{VALUE}{The P/F value to change to}
}
\value{
updated vault
}
\description{
Changes pass/fail value of a single variant based on provided values
}
