% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depracated.R
\name{rescueVariants}
\alias{rescueVariants}
\title{Changes pass/fail value of variants based on provided table}
\usage{
rescueVariants(V, TBL)
}
\arguments{
\item{V}{\itemize{
\item vault
}}

\item{TBL}{a table containing id,var_id,table,&pass value}
}
\value{
updated vault
}
\description{
Changes pass/fail value of variants based on provided table
}
