% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{reviewFeatures}
\alias{reviewFeatures}
\title{compiles data types for manual review of sample-level features
i.e. is a sample SPOP-mutant or WNT-altered}
\usage{
reviewFeatures(MVF, FEATURES)
}
\arguments{
\item{MVF}{filtered metavault}

\item{FEATURES}{vector of features; gene_names}
}
\value{
A table that can be reviewed by hand
}
\description{
compiles data types for manual review of sample-level features
i.e. is a sample SPOP-mutant or WNT-altered
}
