% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistic.R
\name{runGistic}
\alias{runGistic}
\title{Run GISTIC using the docker image at https://github.com/ShixiangWang/install_GISTIC}
\usage{
runGistic(MV, samples = NULL, out.fn, SEG_CN_METHOD = "centered_c")
}
\arguments{
\item{MV}{Metavault containing all the samples}

\item{samples}{samples to include in analysis (if NULL it is all the samples)}

\item{out.fn}{output directory to write the results}

\item{SEG_CN_METHOD}{CNV value to be used for segment table of Gistic}
}
\value{
0
}
\description{
Run GISTIC using the docker image at https://github.com/ShixiangWang/install_GISTIC
}
