% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{setMetaField}
\alias{setMetaField}
\title{Sets metadata field}
\usage{
setMetaField(VAULT, FIELD, VALUE)
}
\arguments{
\item{VAULT}{The vault object}

\item{FIELD}{The metadata field to change}

\item{VALUE}{The value to set the metadata to

@export}
}
\description{
Sets metadata field
}
