% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_rust.R
\name{setup_rust}
\alias{setup_rust}
\title{Setup Rust Toolchain}
\usage{
setup_rust(force = FALSE)
}
\arguments{
\item{force}{If \code{TRUE}, installation proceeds without asking for user
confirmation.}
}
\value{
Invisibly, \code{TRUE} if successful and \code{FALSE} otherwise.
}
\description{
This function downloads the \samp{rustup} installer, run it, and adds targets
to compile for all the CRAN build machines.
}
