% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{setVaultMap}
\alias{setVaultMap}
\title{Sets specific vault map}
\usage{
setVaultMap(V, FIELD, MAP)
}
\arguments{
\item{V}{vault}

\item{FIELD}{which map}

\item{MAP}{new map}
}
\value{
vault
}
\description{
Sets specific vault map
}
