% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{setVaultMeta}
\alias{setVaultMeta}
\title{Sets specific vault meta}
\usage{
setVaultMeta(V, META)
}
\arguments{
\item{V}{vault}

\item{META}{new meta}
}
\value{
vault
}
\description{
Sets specific vault meta
}
