% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{setVaultTable}
\alias{setVaultTable}
\title{Sets specific vault table}
\usage{
setVaultTable(V, FIELD, TBL)
}
\arguments{
\item{V}{vault}

\item{FIELD}{which table}

\item{TBL}{new table}
}
\value{
vault
}
\description{
Sets specific vault table
}
