% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shlib.R
\name{shlib_get}
\alias{shlib_get}
\title{Cache a Shared Library}
\usage{
shlib_get(pkgname)
}
\arguments{
\item{pkgname}{A string giving the name of the package.}
}
\value{
A string giving the path to the cached shared library, or
\code{NULL} if the library is not cached.
}
\description{
This function retrieves the path to a cached shared library developed with
the Cargo framework.
}
\examples{
shlib_get("my_package")

}
\seealso{
shlib_set
}
