% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoprint.R
\name{simplifyOncoMatrixFeature}
\alias{simplifyOncoMatrixFeature}
\title{Simplifies multiple oncomatrix features}
\usage{
simplifyOncoMatrixFeature(OM, NF, OFS, VALUE = NULL)
}
\arguments{
\item{OM}{A prepared OncoMatrix}

\item{NF}{Character value for new feature}

\item{OFS}{vector of old feature names from OM}

\item{VALUE}{(optional) value to set feature as}
}
\value{
A simplified matrix with old features removed and new added
}
\description{
Simplifies multiple oncomatrix features
}
