% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depracated.R
\name{splitMetaVault}
\alias{splitMetaVault}
\title{Creates a list of vaults from a metavault}
\usage{
splitMetaVault(METAVAULT)
}
\arguments{
\item{METAVAULT}{A metaVault}
}
\value{
list of vaults
}
\description{
Creates a list of vaults from a metavault
}
