% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaVault_annotate.R
\name{subsetAnnotatedMetaVault}
\alias{subsetAnnotatedMetaVault}
\title{Given a vector of ids, subsets vault}
\usage{
subsetAnnotatedMetaVault(MVA, IDS)
}
\arguments{
\item{MVA}{An annotated metavault}

\item{IDS}{a vector of IDs}
}
\value{
metavault
}
\description{
Given a vector of ids, subsets vault
}
