% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_annotate.R
\name{tableHitsMetavault}
\alias{tableHitsMetavault}
\title{Counts hits/gene for a single vault}
\usage{
tableHitsMetavault(
  MVF,
  HIT_FIELDS = c("somatic", "germline", "structural", "gene.copy"),
  RULES = NULL,
  GENES = NULL
)
}
\arguments{
\item{MVF}{A vault/metavault}

\item{HIT_FIELDS}{the tables to count towards hits}

\item{RULES}{variant filtering rules}
}
\value{
DF of hits/gene
}
\description{
Counts hits/gene for a single vault
}
