% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargo.R
\name{target}
\alias{target}
\title{Determine the Rust Build Target}
\usage{
target(cran = FALSE)
}
\arguments{
\item{cran}{Are targets for all CRAN build machines desired?}
}
\value{
If \code{cran=FALSE}, a string giving a Rust target, or \code{""} if
this cannot be determined.  If \code{cran=TRUE}, a character vector giving
the targets necessary for CRAN build machines.
}
\description{
This function tries to determine the appropriate Rust target for this
instance of R.  Or, it gives the targets necessary for CRAN build machines.
}
\examples{
target()

}
\seealso{
cross_compile
}
