% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depracated.R
\name{triageVault}
\alias{triageVault}
\title{Adds pass/fail value to vault tables based on annotation, evidence, consequence,
and artifact filters}
\usage{
triageVault(
  vault,
  anno = NULL,
  fset = NULL,
  annotation = TRUE,
  evidence = TRUE,
  consequence = TRUE,
  artifact = TRUE,
  pid = TRUE,
  GOI = NULL
)
}
\arguments{
\item{vault}{vault/metavault}

\item{anno}{anno object (list); contains gene info}

\item{fset}{filter set (name); optional name of the custom filters}

\item{annotation}{boolean; whether filter on annotation}

\item{evidence}{boolean; whether filter on evidence}

\item{consequence}{boolean; whether filter on consequence}

\item{artifact}{boolean; whether filter on artifact}

\item{pid}{boolean; whether to filter on PID}

\item{GOI}{optional vector of genes of interest}
}
\value{
vault w/ pass/fail info for features
}
\description{
Adds pass/fail value to vault tables based on annotation, evidence, consequence,
and artifact filters
}
