% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triage.R
\name{vaultFilter}
\alias{vaultFilter}
\title{Removes variants based on triaging result.}
\usage{
vaultFilter(dxv, rules = NULL)
}
\arguments{
\item{dxv}{dbVault dpVault dtVault}

\item{rules}{variant filtering rules}
}
\value{
dxVault

default rules:
list(
somatic="annotation|consequence|evidence|artifact",
germline="annotation|consequence|evidence|goi",
structural="evidence",
fusion="evidence"
)
}
\description{
Removes variants based on triaging result.
}
