% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triage.R
\name{vaultTriage}
\alias{vaultTriage}
\title{Determines filter status based on annotation, evidence, consequence, artifact, and pid}
\usage{
vaultTriage(
  dxv,
  anno = NULL,
  fset = NULL,
  somatic = TRUE,
  germline = TRUE,
  structural = TRUE,
  fusion = TRUE,
  update = TRUE
)
}
\arguments{
\item{dxv}{dbVault dpVault dtVault}

\item{anno}{custom dtVaultAnno object; optional}

\item{fset}{custom filter set (filename or name); optional}

\item{somatic}{boolean; whether to triage somatic}

\item{germline}{boolean; whether to triage germline}

\item{structural}{boolean; whether to triage structural}

\item{fusion}{boolean; whether to triage fusion}

\item{update}{boolean; update vault tables \link{default: TRUE}}
}
\value{
dxVault or triage result
}
\description{
Determines filter status based on annotation, evidence, consequence, artifact, and pid
}
