% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_documentation.R
\name{api_documentation}
\alias{api_documentation}
\title{Browse API Documentation}
\usage{
api_documentation(pkgroot = ".")
}
\arguments{
\item{pkgroot}{The root directory of the package.}
}
\value{
\code{NULL}, invisibly.
}
\description{
This function opens in a web browser the documentation of the API for the
cargo framework.
}
