% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargo.R
\name{install}
\alias{install}
\title{Install or Update Cargo}
\usage{
install(force = FALSE)
}
\arguments{
\item{force}{Install without asking for user confirmation?}
}
\value{
\code{TRUE} if only only if the installation was successful,
invisibly.
}
\description{
This function downloads and runs the \code{rustup} installer.  If already
installed, the installer simply updates the installation.  Note that,
contrary to the default behavior of the installer, the system is NOT
configured to modify the \code{PATH} environment variable.
}
\examples{
\dontrun{
install()
}

}
