% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_package.R
\name{new_package}
\alias{new_package}
\title{Make a Skeleton for a New Package}
\usage{
new_package(path, ...)
}
\arguments{
\item{path}{A path where the package is created.  The name of the
package is taken as the last element in the file path.}

\item{...}{Extra arguments that are currently ignored.}
}
\description{
A new Rust-based package using the cargo framework is created at the
supplied path and the package is installed.
}
