% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargo.R
\name{target}
\alias{target}
\title{Determine Rust Target}
\usage{
target()
}
\value{
A string giving a Rust target
}
\description{
This function determines the appropriate Rust target for this instance of R.
}
\examples{
tryCatch(target(), error=function(e) cat(e$message, "\n", sep=""))

}
