% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargo.R
\name{target}
\alias{target}
\title{Determine the Rust Build Target}
\usage{
target()
}
\value{
A string giving a Rust target, or \code{""} if this cannot be
determined.
}
\description{
This function tries to determine the appropriate Rust target for this
instance of R.
}
\examples{
target()

}
